When building values from map, check for null.
In the RedisMapState, buildValuesFromMap takes a list of keys and a map of key/value pairs (as returned by multiGet), returning a List of Strings as the value for the key. If the multiGet returns a null, an error will be thrown when calling `new String(bytes)` where bytes is null. This commit alters buildValuesFromMap to mirror the logic used in deserializeValues.